clever-crypt/requirements.txt
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.0
PyMySQL==1.0.2
cryptography==35.0.0